projects/cymbal-be-TH/src/llm_cr_be_requirements.txt
fastapi==0.101.0
hypercorn==0.14.4
langchain>=0.0.263
unstructured>=0.9.2
pydantic==1.10.12
shapely==1.8.5
faiss-cpu==1.7.4
google-cloud-aiplatform>=1.30.1
google-cloud-translate>=3.12.0
google-auth>=2.22.0
google-cloud-logging>=3.6.0
tabulate>=0.9.0
iso639>=0.1.4
configparser>=6.0.0
pytz